refractor code to use event driven pattern by creating spearte files to handle his responbility as expected
