+++ backend/src/users/controller.ts
@@ -1,0 +1,52 @@
+import { Request, Response } from "express";
+import { GetCommand, PutCommand } from "@aws-sdk/lib-dynamodb";
+import { docClient } from "../utils/dynamoClient";
+import { User } from "./model";
+
+const USERS_TABLE = process.env.USERS_TABLE;
+
+export const getUser = async (req: Request, res: Response) => {
+  const params = {
+    TableName: USERS_TABLE,
+    Key: {
+      userId: req.params.userId,
+    },
+  };
+
+  try {
+    const command = new GetCommand(params);
+    const { Item } = await docClient.send(command);
+    if (Item) {
+      const { userId, name } = Item as User;
+      res.json({ userId, name });
+    } else {
+      res.status(404).json({ error: 'Could not find user with provided "userId"' });
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({ error: "Could not retrieve user" });
+  }
+};
+
+export const createUser = async (req: Request, res: Response) => {
+  const { userId, name } = req.body as User;
+  if (typeof userId !== "string") {
+    return res.status(400).json({ error: '"userId" must be a string' });
+  } else if (typeof name !== "string") {
+    return res.status(400).json({ error: '"name" must be a string' });
+  }
+
+  const params = {
+    TableName: USERS_TABLE,
+    Item: { userId, name },
+  };
+
+  try {
+    const command = new PutCommand(params);
+    await docClient.send(command);
+    res.json({ userId, name });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ error: "Could not create user" });
+  }
+};
