+++ backend/src/users/handler.ts
@@ -1,0 +1,16 @@
+import express from "express";
+import serverless from "serverless-http";
+import { getUser, createUser } from "./controller";
+
+const app = express();
+
+app.use(express.json());
+
+app.get("/users/:userId", getUser);
+app.post("/users", createUser);
+
+app.use((req, res) => {
+  res.status(404).json({ error: "Not Found" });
+});
+
+export const handler = serverless(app);
