+++ backend/src/users/model.ts
@@ -1,0 +1,5 @@
+export interface User {
+    userId: string;
+    name: string;
+  }
+
