+++ backend/src/users/service.ts
@@ -1,0 +1,1 @@
+// Define any additional business logic here that might be shared across multiple controllers or used internally
