+++ backend/src/utils/dynamoClient.ts
@@ -1,0 +1,5 @@
+import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
+import { DynamoDBDocumentClient } from "@aws-sdk/lib-dynamodb";
+
+const client = new DynamoDBClient({});
+export const docClient = DynamoDBDocumentClient.from(client);
